requirements.txt
pyautogui
opencv-python
pillow
keyboard
numpy
mss
